Preparata la struttura del caso d'uso
