--- app/src/main/java/it/unimib/brewday/ui/ListaRicetteFragment.java
@@ -1,32 +1,0 @@
-package it.unimib.brewday.ui;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import it.unimib.brewday.R;
-
-public class ListaRicetteFragment extends Fragment {
-
-
-    public ListaRicetteFragment() {
-        // Required empty public constructor
-    }
-
-    public static ListaRicetteFragment newInstance() {
-
-        return new ListaRicetteFragment();
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_lista_ricette, container, false);
-    }
-}
