+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/listaRicetteFragment"
-        android:name="it.unimib.brewday.ui.ListaRicetteFragment"
+        android:name="it.unimib.brewday.ui.gestisci_ricette.ListaRicetteFragment"
         android:label="fragment_lista_ricette"
         tools:layout="@layout/fragment_lista_ricette" />
     <fragment
@@ -17,7 +17,7 @@
         tools:layout="@layout/fragment_birre" />
     <fragment
         android:id="@+id/creaRicettaFragment"
-        android:name="it.unimib.brewday.ui.CreaRicettaFragment"
+        android:name="it.unimib.brewday.ui.gestisci_ricette.CreaRicettaFragment"
         android:label="fragment_crea_ricetta"
         tools:layout="@layout/fragment_crea_ricetta" />
     <fragment
